Remove `GNUInstallDirs` module
The `GNUInstallDirs` module defines various `CMAKE_INSTALL_<dir>`
variables that can be used as the destinations for the CMake `install`
directives following the GNU convention. However, the GNU convention is
to install 64-bit shared objects (.so) in the `/usr/lib64` folder when
the architecture is 64-bit, but when compiling the shared object on
a properly configured cRIO, it is expected to be installed in the
`/usr/lib` directory. This causes the Call Library Node function in
LabVIEW to fail because the `libssh2lv` shared object is installed in
`/usr/lib64` and its dependencies (libssh2) are expected to be in the
same folder as it.

Thus, the GNUInstallDirs module is removed and replaced with hard-coded
destinations for the installation of the shared object, symbolic links,
and headers (`/usr/include`).