Fix package version number
There is already a v0.4 tag in the repository, but the project/package
version in the CMake configuration file (CMakeLists.txt) is only at
v0.3. The project/package version is bumped to v0.5 in preparation for
the next release.

I think I was trying to keep the project/package version the same for
each "release" and only change it if a "significant" change to _the
package_ occurred. But this is just confusing and it is easier to just
bump the minor version number for each release for the package,
especially after being away from the project for so long.